tests/dialog-autofocus-multiple-times.html

<!DOCTYPE html>
<html>
<head>
<script src="./dist/polyfill.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/common.js"></script>
<script>
promise_test(function() {
  return waitUntilLoadedAndAutofocused().then(function() {
        // 2020-09-01 - Manually focused until Webkit autofocus behaviour is
        // resolved: https://bugs.webkit.org/show_bug.cgi?id=203139
        if (document.activeElement !== document.getElementById("outer-button")) {
            document.getElementById("outer-button").focus();
        }
        assert_equals(document.activeElement, document.getElementById("outer-button"));

        var focusCount = 0;
        var dlg = document.getElementById("dlg");
        var input1 = document.getElementById("input1");
        var input2 = document.getElementById("input2");
        input2.onfocus = function() { focusCount += 1 };

        var expectedFocusCount = 3;
        for (i = 0; i < expectedFocusCount; i++) {
            dlg.show();
            assert_equals(document.activeElement, input2);
            input1.focus();
            assert_equals(document.activeElement,input1);
            dlg.close();
        }

        assert_equals(focusCount.toString(), expectedFocusCount.toString());
  });
}, "autofocus is run every time a dialog is opened");
</script>
</head>
<body>
<button id="outer-button" autofocus></button>
<dialog id="dlg">
    <!-- Unfocusable elements with [autofocus] should be ignored. -->
    <input autofocus disabled>
    <textarea autofocus hidden></textarea>
    <input id="input1"></input>
    <input id="input2" autofocus></input>
</dialog>
</body>
</html>
